[UPLOAD] lista de preguntas mostrada con spinner

+++ evaluacion_docente/src/app/lista-preguntas/preguntas-list/preguntas-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { PreguntaService } from '../../services/preguntas.service';
-import { ModalController } from '@ionic/angular';
+import { ModalController, NavController } from '@ionic/angular';
 import { AgregarPreguntaPage } from '../../agregar-pregunta/agregar-pregunta.page';
 import { ModificarPreguntaPage } from '../../modificar-pregunta/modificar-pregunta.page';
 
@@ -10,45 +10,58 @@ import { ModificarPreguntaPage } from '../../modificar-pregunta/modificar-pregun
   styleUrls: ['./preguntas-list.component.scss']
 })
 export class ListaPreguntasComponent implements OnInit {
-  preguntas: string[] = [];
+  preguntas: { texto: string, calificacion: number }[] = [];
   esAdmin: boolean = false; // Valor por defecto
 
   constructor(
     private preguntaService: PreguntaService,
-    private modalCtrl: ModalController
+    private modalCtrl: ModalController,
+    private navCtrl: NavController
   ) {}
 
   ngOnInit() {
     this.preguntas = this.preguntaService.preguntas;
-    // Obtener el estado de administrador desde algún servicio o lógica de autenticación
     this.esAdmin = this.obtenerEstadoAdmin();
   }
 
   obtenerEstadoAdmin(): boolean {
-    // Lógica para determinar si el usuario es administrador
-    // Ejemplo: leer de localStorage o un servicio de autenticación
-    return true; // Cambiar a la lógica real
+    return localStorage.getItem('isAdmin') === 'true'; // Leer el estado de admin desde el almacenamiento local
+  }
+
+  goToListaProfesores() {
+    this.navCtrl.navigateForward('/materias');
   }
 
   async abrirModalAgregarPregunta() {
-    const modal = await this.modalCtrl.create({
-      component: AgregarPreguntaPage,
-    });
-    await modal.present();
+    if (this.esAdmin) {
+      const modal = await this.modalCtrl.create({
+        component: AgregarPreguntaPage,
+      });
+      await modal.present();
+    }
   }
 
   async abrirModalModificarPregunta(pregunta: string, index: number) {
-    const modal = await this.modalCtrl.create({
-      component: ModificarPreguntaPage,
-      componentProps: {
-        pregunta,
-        index
-      }
-    });
-    await modal.present();
+    if (this.esAdmin) {
+      const modal = await this.modalCtrl.create({
+        component: ModificarPreguntaPage,
+        componentProps: {
+          pregunta,
+          index
+        }
+      });
+      await modal.present();
+    }
   }
 
   eliminarPregunta(index: number) {
-    this.preguntaService.eliminarPregunta(index);
+    if (this.esAdmin) {
+      this.preguntaService.eliminarPregunta(index);
+    }
+  }
+
+  actualizarCalificacion(index: number, event: any) {
+    const calificacion = event.target.value;
+    this.preguntaService.modificarCalificacion(index, calificacion);
   }
 }
